+++ tests/utils.py
@@ -329,8 +329,8 @@ def get_test_webhook_app(webhook_url):
         user = body["user_name"]
         # Since we can't call a local magpie app directly here, we save the tmp_url here,
         # and retrieve it in the test case
-        settings["tmp_url"] = body["tmp_url"]
-        return Response("Failing webhook url with user " + user + " and tmp_url " + settings["tmp_url"])
+        settings["callback_url"] = body["callback_url"]
+        return Response("Failing webhook url with user " + user + " and callback_url " + settings["callback_url"])
 
     def get_status(*_):
         # Returns the status number
@@ -338,7 +338,7 @@ def get_test_webhook_app(webhook_url):
 
     def get_tmp_url(*_):
         # Returns the tmp_url
-        return Response(str(settings["tmp_url"]))
+        return Response(str(settings["callback_url"]))
 
     def check_payload(request):
         # Check if the input payload is present in the webhook app saved payload
@@ -348,7 +348,7 @@ def get_test_webhook_app(webhook_url):
     def reset(*_):
         settings["webhook_status"] = 0
         settings["payload"] = []
-        settings["tmp_url"] = ""
+        settings["callback_url"] = ""
         return Response("Webhook app has been reset.")
 
     with Configurator() as config:
@@ -356,12 +356,12 @@ def get_test_webhook_app(webhook_url):
         # Initialize status
         settings["webhook_status"] = 0
         settings["payload"] = []
-        settings["tmp_url"] = ""
+        settings["callback_url"] = ""
         config.add_route("webhook_create", "/webhook_create")
         config.add_route("webhook_delete", "/webhook_delete")
         config.add_route("webhook_fail", "/webhook_fail")
         config.add_route("get_status", "/get_status")
-        config.add_route("get_tmp_url", "/get_tmp_url")
+        config.add_route("get_callback_url", "/get_callback_url")
         config.add_route("check_payload", "/check_payload")
         config.add_route("reset", "/reset")
         config.add_view(webhook_create_request, route_name="webhook_create",
@@ -371,7 +371,7 @@ def get_test_webhook_app(webhook_url):
         config.add_view(webhook_fail_request, route_name="webhook_fail",
                         request_method="POST")
         config.add_view(get_status, route_name="get_status", request_method="GET")
-        config.add_view(get_tmp_url, route_name="get_tmp_url", request_method="GET")
+        config.add_view(get_tmp_url, route_name="get_callback_url", request_method="GET")
         config.add_view(check_payload, route_name="check_payload", request_method="POST")
         config.add_view(reset, route_name="reset", request_method="POST")
         webhook_app_instance = config.make_wsgi_app()